[Chore] #47 - DropDown 디테일 반영 및 네이밍 변경
